SHOPDEV-36 Fix RC releases

+++ EnvironmentServer.Daemon/Actions/CloneProductionTemplate.cs
@@ -22,14 +22,24 @@ public class CloneProductionTemplate : ActionBase
         var version = File.ReadAllText($"/home/{user.Username}/files/{env.InternalName}/version.txt");
         File.Delete($"/home/{user.Username}/files/{env.InternalName}/version.txt");
 
-        await Bash.CommandAsync($"git clone --branch v{version} https://github.com/shopware/production.git {homeDir}", homeDir);
+        if (version.ToLower().Contains("rc"))
+        {
+            await Bash.CommandAsync($"git clone --branch v{version} https://github.com/shopware/plattform.git {homeDir}", homeDir);
+            await Bash.CommandAsync($"composer setup -q", homeDir, validation: false);
+        }
+        else
+        {
+            await Bash.CommandAsync($"git clone --branch v{version} https://github.com/shopware/production.git {homeDir}", homeDir);
+        }        
 
         await Bash.CommandAsync($"composer install -q", homeDir, validation: false);
 
         File.Delete($"{homeDir}/vendor/shopware/recovery/composer.lock");
         File.Delete($"{homeDir}/vendor/shopware/recovery/Common/composer.lock");
 
-        File.Move($"{homeDir}/public/.htaccess.dist", $"{homeDir}/public/.htaccess");
+        if (File.Exists($"{homeDir}/public/.htaccess.dist"))
+            File.Move($"{homeDir}/public/.htaccess.dist", $"{homeDir}/public/.htaccess");
+
 
         await Bash.CommandAsync($"bin/console assets:install", homeDir, validation: false);
